Move PRIVEC output read section of main function to privec_reader
